backend/requirements-vercel.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
motor==3.3.2
pymongo==4.6.0

# HTTP client
httpx==0.25.2

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Email
aiofiles==23.2.1

# Environment
python-decouple==3.8

# Async support
asyncio-mqtt==0.13.0